pipeline/assets/ingestion/requirements.txt
pandas==2.2.0
requests==2.31.0

google-cloud-bigquery
